Move SaveArticle logic to unidirectional.
