+++ app/src/main/java/com/greglaun/lector/store/sideeffect/persistence/PersistenceSideEffect.kt
@@ -1,19 +1,45 @@
 package com.greglaun.lector.store.sideeffect.persistence
 
-import com.greglaun.lector.store.Action
-import com.greglaun.lector.store.SideEffect
-import com.greglaun.lector.store.Store
+import com.greglaun.lector.data.cache.ResponseSource
+import com.greglaun.lector.data.course.CourseDownloader
+import com.greglaun.lector.data.course.CourseSource
+import com.greglaun.lector.store.*
+import com.greglaun.lector.ui.speak.ArticleStateSource
+import com.greglaun.lector.ui.speak.JSoupArticleStateSource
 
-class PersistenceSideEffect(val store: Store)
-    : SideEffect {
+class PersistenceSideEffect(val store: Store, val responseSource: ResponseSource,
+                            val courseSource: CourseSource,
+                            val courseDownloader: CourseDownloader,
+                            val articleStateSource: ArticleStateSource = JSoupArticleStateSource(
+                              responseSource)): SideEffect {
 
 
     override suspend fun handle(action: Action) {
-//        when (action) {
-//            is CreationAction -> CreationHandler.handle(action) { store.dispatch(it) }
-//            is UpdateAction -> UpdateHandler.handle(action) { store.dispatch(it) }
-//            is ReadAction -> ReadHandler.handle(action) { store.dispatch(it) }
-//            is DeleteAction -> DeleteHandler.handle(action) { store.dispatch(it) }
-//        }
+        when (action) {
+            is ReadAction.FetchCourseDetailsAction ->
+                handleFetchCourseDetails(action, courseDownloader) { store.dispatch(it) }
+            is ReadAction.LoadNewUrlAction -> loadNewUrl(
+                    action,
+                    responseSource,
+                    articleStateSource)  {
+                store.dispatch(it)
+            }
+            is ReadAction.FetchAllPermanentAndDisplay ->
+                handleFetchAllPermanentAndDisplay(responseSource) {
+                    store.dispatch(it)
+                }
+            is ReadAction.FetchCourseInfoAndDisplay ->
+                handleFetchCourseInfoAndDisplay(action, courseSource) {
+                    store.dispatch(it)
+                }
+            is UpdateAction.ArticleOverAction -> handleArticleOver(
+                    store,
+                    responseSource,
+                    courseSource,
+                    articleStateSource) {
+                store.dispatch(it)
+            }
+            is WriteAction.SaveArticle -> handleSaveArticle(action, responseSource)
+        }
     }
 }
